Refactored machine type data conversions
